changes: zip json before storage, add mappers, fix tests

+++ src/main/java/se/sundsvall/educationdata/service/PlannedEducationService.java
@@ -1,12 +1,12 @@
 package se.sundsvall.educationdata.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.ArrayList;
+
 import java.util.List;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import se.sundsvall.dept44.problem.Problem;
 import se.sundsvall.educationdata.integration.db.ReferenceCategoryRepository;
 import se.sundsvall.educationdata.integration.db.model.ReferenceCategory;
 import se.sundsvall.educationdata.integration.plannededucation.PlannedEducationIntegration;
@@ -16,36 +16,21 @@ public class PlannedEducationService {
 
 	private final PlannedEducationIntegration integration;
 	private final ReferenceCategoryRepository referenceCategoryRepository;
-	private final ObjectMapper objectMapper;
 
-	public PlannedEducationService(PlannedEducationIntegration integration, ReferenceCategoryRepository referenceCategoryRepository, ObjectMapper objectMapper) {
+	public PlannedEducationService(PlannedEducationIntegration integration, ReferenceCategoryRepository referenceCategoryRepository) {
 		this.integration = integration;
 		this.referenceCategoryRepository = referenceCategoryRepository;
-		this.objectMapper = objectMapper;
 	}
 
-	public ResponseEntity<Void> getEducationInfo() throws JsonProcessingException {
-		String json = integration.getAllAreas();
+	@Transactional
+	public void getCategoryInfo() throws JsonProcessingException {
+		List<ReferenceCategory> categories = integration.getAllAreas();
 
-		JsonNode areas = objectMapper.readTree(json)
-			.path("body")
-			.path("areas");
-		List<ReferenceCategory> rows = new ArrayList<>();
-		for (JsonNode area : areas) {
-			String categoryId = area.path("areaId").asText();
-			String categoryName = area.path("name").asText();
-
-			for (JsonNode direction : area.path("directions")) {
-				rows.add(ReferenceCategory.builder()
-					.withCategoryId(categoryId)
-					.withCategoryName(categoryName)
-					.withDirectionId(direction.path("directionId").asText())
-					.withDirectionName(direction.path("name").asText()).build());
-			}
-		}
+		if (categories.isEmpty())
+			throw Problem.valueOf(HttpStatus.BAD_GATEWAY,
+				"No content");
 
 		referenceCategoryRepository.deleteAllInBatch();
-		referenceCategoryRepository.saveAll(rows);
-		return ResponseEntity.ok().build();
+		referenceCategoryRepository.saveAll(categories);
 	}
 }
